requirements/requirements.txt
alchemy-catalyst
catalyst==20.5
transformers>=1.9
reaction
torch_optimizer
wandb
